:truck: [#2471] -- move test modules into their respective contrib package
They test the specific plugins, so the tests should live in the plugin
package.
